refactor(csp): decouple central CSP builder from optional provider modules (#364)
scripts/csp.ts statically imported five commerce/booking provider helper
modules (snipcart, shopify-buy-button, paddle, lemon-squeezy, booking) just
to call their build*CSP() helpers. That made the modules load-bearing for
the module graph even on content-only sites with no checkout or booking —
they couldn't be deleted without also hand-editing csp.ts.

Centralize the five CSP builders into csp.ts alongside the ones that already
live there (buildPolarCSP, buildTurnstileCSP, buildGiscusCSP, …) and drop the
static imports. BookingProvider is now defined locally in csp.ts so there's
no type dependency on booking.ts either. CSP output is byte-for-byte
unchanged; buildCSP stays synchronous (no async ripple to its ~30 callers).

Now nothing in the site build graph imports the provider helper modules, so a
content-only site can delete snipcart.ts/shopify-buy-button.ts/paddle.ts/
lemon-squeezy.ts/booking.ts without touching csp.ts.

- Removed build*CSP + their CSP interfaces from the five provider modules
  (booking.ts keeps the BookingProvider type, still used by its embed helpers).
- Moved the builder unit tests into csp.test.ts; added a guard test asserting
  csp.ts never re-imports a provider module.
- Repointed skills/booking/SKILL.md to csp.ts for buildBookingCSP().
